+++ include/rocksdb/env.h
@@ -363,7 +363,7 @@ class Env : public Customizable {
   virtual Status DeleteFile(const std::string& fname) = 0;
 
   // Truncate the named file to the specified size.
-  virtual Status Truncate(const std::string& /*fname*/, size_t /*size*/) {
+  virtual Status Truncate(const std::string& /*fname*/, uint64_t /*size*/) {
     return Status::NotSupported("Truncate is not supported for this Env");
   }
 
@@ -725,7 +725,7 @@ class SequentialFile {
   // Remove any kind of caching of data from the offset to offset+length
   // of this file. If the length is 0, then it refers to the end of file.
   // If the system is not caching the file contents, then this is a noop.
-  virtual Status InvalidateCache(size_t /*offset*/, size_t /*length*/) {
+  virtual Status InvalidateCache(uint64_t /*offset*/, uint64_t /*length*/) {
     return Status::NotSupported(
         "SequentialFile::InvalidateCache not supported.");
   }
@@ -842,7 +842,7 @@ class RandomAccessFile {
   // Remove any kind of caching of data from the offset to offset+length
   // of this file. If the length is 0, then it refers to the end of file.
   // If the system is not caching the file contents, then this is a noop.
-  virtual Status InvalidateCache(size_t /*offset*/, size_t /*length*/) {
+  virtual Status InvalidateCache(uint64_t /*offset*/, uint64_t /*length*/) {
     return Status::NotSupported(
         "RandomAccessFile::InvalidateCache not supported.");
   }
@@ -1002,7 +1002,7 @@ class WritableFile {
   // of this file. If the length is 0, then it refers to the end of file.
   // If the system is not caching the file contents, then this is a noop.
   // This call has no effect on dirty pages in the cache.
-  virtual Status InvalidateCache(size_t /*offset*/, size_t /*length*/) {
+  virtual Status InvalidateCache(uint64_t /*offset*/, uint64_t /*length*/) {
     return Status::NotSupported("WritableFile::InvalidateCache not supported.");
   }
 
@@ -1024,7 +1024,7 @@ class WritableFile {
   // of space on devices where it can result in less file
   // fragmentation and/or less waste from over-zealous filesystem
   // pre-allocation.
-  virtual void PrepareWrite(size_t offset, size_t len) {
+  virtual void PrepareWrite(uint64_t offset, size_t len) {
     if (preallocation_block_size_ == 0) {
       return;
     }
@@ -1166,7 +1166,7 @@ enum InfoLogLevel : unsigned char {
 // including data loss, unreported corruption, deadlocks, and more.
 class Logger {
  public:
-  size_t kDoNotSupportGetLogFileSize = (std::numeric_limits<size_t>::max)();
+  uint64_t kDoNotSupportGetLogFileSize = (std::numeric_limits<uint64_t>::max)();
 
   explicit Logger(const InfoLogLevel log_level = InfoLogLevel::INFO_LEVEL)
       : closed_(false), log_level_(log_level) {}
@@ -1206,7 +1206,9 @@ class Logger {
   virtual void Logv(const InfoLogLevel log_level, const char* format,
                     va_list ap);
 
-  virtual size_t GetLogFileSize() const { return kDoNotSupportGetLogFileSize; }
+  virtual uint64_t GetLogFileSize() const {
+    return kDoNotSupportGetLogFileSize;
+  }
   // Flush to the OS buffers
   virtual void Flush() {}
   virtual InfoLogLevel GetInfoLogLevel() const { return log_level_; }
@@ -1453,7 +1455,7 @@ class EnvWrapper : public Env {
   Status DeleteFile(const std::string& f) override {
     return target_.env->DeleteFile(f);
   }
-  Status Truncate(const std::string& fname, size_t size) override {
+  Status Truncate(const std::string& fname, uint64_t size) override {
     return target_.env->Truncate(fname, size);
   }
   Status CreateDir(const std::string& d) override {
@@ -1651,7 +1653,7 @@ class SequentialFileWrapper : public SequentialFile {
   size_t GetRequiredBufferAlignment() const override {
     return target_->GetRequiredBufferAlignment();
   }
-  Status InvalidateCache(size_t offset, size_t length) override {
+  Status InvalidateCache(uint64_t offset, uint64_t length) override {
     return target_->InvalidateCache(offset, length);
   }
   Status PositionedRead(uint64_t offset, size_t n, Slice* result,
@@ -1686,7 +1688,7 @@ class RandomAccessFileWrapper : public RandomAccessFile {
   size_t GetRequiredBufferAlignment() const override {
     return target_->GetRequiredBufferAlignment();
   }
-  Status InvalidateCache(size_t offset, size_t length) override {
+  Status InvalidateCache(uint64_t offset, uint64_t length) override {
     return target_->InvalidateCache(offset, length);
   }
 
@@ -1753,7 +1755,7 @@ class WritableFileWrapper : public WritableFile {
     return target_->GetUniqueId(id, max_size);
   }
 
-  Status InvalidateCache(size_t offset, size_t length) override {
+  Status InvalidateCache(uint64_t offset, uint64_t length) override {
     return target_->InvalidateCache(offset, length);
   }
 
@@ -1761,7 +1763,7 @@ class WritableFileWrapper : public WritableFile {
     return target_->RangeSync(offset, nbytes);
   }
 
-  void PrepareWrite(size_t offset, size_t len) override {
+  void PrepareWrite(uint64_t offset, size_t len) override {
     target_->PrepareWrite(offset, len);
   }
 
@@ -1825,7 +1827,7 @@ class LoggerWrapper : public Logger {
             va_list ap) override {
     return target_->Logv(log_level, format, ap);
   }
-  size_t GetLogFileSize() const override { return target_->GetLogFileSize(); }
+  uint64_t GetLogFileSize() const override { return target_->GetLogFileSize(); }
   void Flush() override { return target_->Flush(); }
   InfoLogLevel GetInfoLogLevel() const override {
     return target_->GetInfoLogLevel();
